# swagger.yaml
swagger: '2.0'
info:
  description: System to keep track of customer orders
  version: 1.0.0
  title: Customer Orders API
  contact:
    email: SwingGuy1024@yahoo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /menuItem:
    put:
      summary: Add a MenuItem, with optional MenuItemOptions.
      description: |
        Add a MenuItem, complete with MenuItemOptions, to the database. Since MenuItemOptions are always linked to specific MenuItems, they are created, updated, and removed by using the MenuItem API. More options may be added later with the /MenuItem/addOption/ API.
      operationId: addMenuItem
      parameters:
        - in: body
          name: menuItem
          description: Complete menu item, with MenuItemOptions
          required: true
          schema:
            $ref: '#/definitions/MenuItemDto'
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: Menu Item Created
          schema:
            $ref: '#/definitions/CreatedResponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/CreatedResponse'
    get:
      summary: Get all menu items.
      description: |
        Retrieve all menu items
      operationId: getAll
      produces:
        - application/json
      responses:
        200:
          description: Found
          schema:
            type: array
            items:
              $ref: '#/definitions/MenuItemDto'
  /menuItem/{id}:
    get:
      summary: Gets a menuItem by ID
      description: Gets a MenuItem by its ID.
      operationId: getMenuItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id of menuItem to find
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: Found
          schema:
            $ref: '#/definitions/MenuItemDto'
        404:
          description: NotFound
  /menuItem/addOption/{menuItemId}:
    post:
      summary: Add a menuItemOption
      description: |
        Add a MenuItemOption to a MenuItem.
      operationId: addMenuItemOption
      parameters:
        - in: path
          name: menuItemId
          required: true
          description: ID of the MenuItem getting the new option
          type: integer
          format: int32
        - in: body
          required: true
          name: option
          description: MenuItemOption
          schema:
            $ref: '#/definitions/MenuItemOptionDto'
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        201:
          description: MenuItemOption added
          schema:
            $ref: '#/definitions/CreatedResponse'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/CreatedResponse'
  /menuItem/deleteOption/{optionId}:
    delete:
      summary: Delete a MenuItemOption
      description: Delete a MenuItemOption from a MenuItem.
      operationId: deleteOption
      parameters:
        - in: path
          name: optionId
          required: true
          description: ID of the MenuItemOpton to delete
          type: integer
          format: int32
      produces:
        - application/json
      responses:
        200:
          description: MenuItemOption deleted
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/CreatedResponse'
...
definitions:
  MenuItemOptionDto:
    type: object
    description: Option for a MenuItem
    properties:
      name:
        type: string
      deltaPrice:
        type: number
        format: int32
        description: Floating point price. Strings are easier to work with.
      id:
        type: integer
        format: int32
  MenuItemDto:
    type: object
    properties:
      name:
        type: string
      itemPrice:
        type: string
        description: Floating point price. Strings are easier to work with.
      allowedOptions:
        type: array
        items:
          $ref: '#/definitions/MenuItemOptionDto'
      id:
        type: integer
        format: int32
    required:
      - name
      - itemPrice
...
  CreatedResponse:
    # This type is primarily to return the id of the 
    # created object.
    properties:
      id:
        type: integer
        format: int32
        # Note on numbers:
        #   integer int32     gives Integer
        #   integer int64     gives Long
        #   number  int32     gives BigDecimal
        #   number  int64     gives BigDecimal
        #   number  double    gives Double
        #   number  float     gives Float
        #   string            gives String
        #   string  byte      gives byte[ ]
        #   string  binary    gives File
        #   string  date      gives LocalDate
        #   string  date-time gives OffsetDateTime
        #   string  password  gives String
        #   boolean           gives Boolean
        # Undocumented:
        #   integer byte      gives Integer
        #   integer binary    gives Integer
        #   integer date      gives Integer
        #   string  int32     gives String
        #   string  int64     gives String
        # I don't know how to get BigInteger.
      # intx:
      #   type: integer
      #   format: decimal
      # bd:
      #   type: integer
      #   format: biginteger
      # bd2:
      #   type: integer
      #   format: bigInteger
      # bd3:
      #   type: integer
      #   format: BigInteger
      # dbl:
      #   type: number
      #   format: double
      # flt:
      #   type: number
      #   format: float
      # str:
      #   type: string
      # pw:
      #   type: string
      #   format: password
      # bool:
      #   type: boolean
      # intByte:
      #   type: integer
      #   format: byte
      # intBinary:
      #   type: integer
      #   format: binary
      # intDate:
      #   type: integer
      #   format: date
      # int64:
      #   type: integer
      #   format: int64
      # string64:
      #   type: string
      #   format: int64
      # numBigInt:
      #   type: number
      #   format: bigint
      # num64:
      #   type: number
      #   format: int64
      message:
        type: string
      body:
        description: Json description of result.
        type: string
    required:
      - id
      - message
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
schemes:
  - https
  - http
# Added by API Auto Mocking Plugin
basePath: /NeptuneDreams/FoodOrders/1.0.0